move all collections-related tests to package "integration.collections"
