+++ src/test/java/integration/collections/NamedCollectionsTest.java
@@ -1,7 +1,8 @@
-package integration;
+package integration.collections;
 
 import com.codeborne.selenide.ElementsCollection;
 import com.codeborne.selenide.ex.ListSizeMismatch;
+import integration.ITest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.By;
